backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiofiles==23.2.1
setuptools
openai-whisper
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.2.1
httpx==0.27.0
pillow==10.2.0
torch==2.2.0
torchaudio==2.2.0
transformers==4.38.2
